RDS Query/amzapi_function/requirements.txt
pybase64
pandas
pymysql
pycryptodome
SQLAlchemy
numpy
python-dotenv
pymysql
python-amazon-ad-api